extract DB Buddy
