+++ spec/spec_helper.rb
@@ -1,4 +1,5 @@
 Dir[File.join(File.dirname(__FILE__), "../lib", "*.rb")].each {|f| require f}
+require 'support/db_buddy'
 
 RSpec.configure do |config|
   # Use color in STDOUT
@@ -10,25 +11,8 @@ RSpec.configure do |config|
   # Use the specified formatter
   config.formatter = :documentation # :progress, :html, :textmate
 
+  #you can do global before/after here like this:
+  #config.before(:each) do
+  #  #code
+  #end
 end
-
-
-
-class DBBuddy
-  #from logan
-  def self.create
-    db = SQLite3::Database.new('students.db')
-    create = "CREATE TABLE IF NOT EXISTS students(id INTEGER PRIMARY KEY AUTOINCREMENT, name TEXT)"
-    db.execute(create)
-    db
-  end
-
-  def self.destroy(db)
-    db.execute("DELETE FROM students")
-    db.execute("DELETE FROM sqlite_sequence WHERE name = 'students'")
-    db.close
-    db
-  end
-
-end
-
